store: Tentatively optimize recordFilteringReader
This commit changes `recordFilteringReader` in an attempt to improve its
performance by specialising it with the current record filtering logic
(less CPU) and replacing the internal calls of `bufio.Scanner.Scan+Bytes`
with `bufio.Reader.ReadSlice` (less CPU and allocations).

The actual benefits from this change still need to be verified with some
form of profiling.

In addition to these optimisations, this commit also fixes a bug where
the query was being applied to the whole record instead of being limited
to the opaque record body. This is now tested in `TestRecordFilteringReader`.