pkg/store: query: allow &{from,to} as ULID
This enables a sort of pagination. Users can make a query from e.g. now,
and cancel the response whenever they get enough data to fulfill a
"page". Then, they can pass the last received ULID as the &from of the
next query, and repeat the process.

A better implementation would allow a &limit parameter, but that's
tricky now, as the lazy-evaluated query reader pipeline doesn't have a
straightforward way to count records. That requires a bit of
refactoring, which requires a bit of benchmarking to make sure we don't
regress on performance.

This should allow pagination in the UI to proceed.