Add browser-based user interface
This change embedds a web interface under /ui for the ingestore and store
processes. Based on the assumption that such an tool will help with the
overall usability of the system.

For the implementation of the client Elm was choosen, which is compiled
to Javascript. All static assets including the HTML and CSS are embedded
in the ui package with the help `esc`. We can avoid scenarios where we
have to ship more than one file and don't break the out-of-the-box
experience.