pkg/store: Size before passing ownership
newConcurrentFilteringReadCloser takes ownership of the passed io.ReadCloser (an
os.File, via fs.realFile) and launches a goroutine that can Close it, in certain
error conditions. If this happens before the calling context gets a chance to
take the realFile.Size, we'll get a panic (use of closed file). So, it's
important that the calling context perform no operations on the passed
io.ReadCloser after passing it.