pkg/store: segments may disappear during Query
A Query can be split into two phases: segment file selection, based on
time range; and reader construction, where we actually open the files.
It can happen that a file is removed (e.g. compacted) between the first
and second phase. In this case, we'll get partial results, but we
shouldn't fatal-out.

A better solution would be to open file handles when we select segments,
rather than just returning file names. I'll try to do that next.