pkg/store: fix bug re: recovering active files
.active files are the default extension for WriteSegments created by the
FileLog. They get UUID filenames because we don't yet know their low and
high ULIDs. Most of the time that's fine: code that expects to parse
filenames for ULID ranges (specifically: query code) knows to ignore
.active files.

When a FileLog is created, it first does a sweep of the underlying
filesystem, and any apparently temporary or intermediate files that are
discovered get recovered (essentially, renamed) to proper .flushed
files. This is a form of crash recovery. There was a bug in this code
that blindly changed the extension of .active files to .flushed, without
also giving them a proper low-high ULID filename. This causes a parse
failure, when query code tries to interpret the UUID as a pair of ULIDs.

This change fixes that, by extracting low and high ULIDs from .active
files, and renaming them properly during recovery.