Bug fixes (#85)
* pkg/store: Size before passing ownership

newConcurrentFilteringReadCloser takes ownership of the passed io.ReadCloser (an
os.File, via fs.realFile) and launches a goroutine that can Close it, in certain
error conditions. If this happens before the calling context gets a chance to
take the realFile.Size, we'll get a panic (use of closed file). So, it's
important that the calling context perform no operations on the passed
io.ReadCloser after passing it.

* pkg/store: segments may disappear during Query

A Query can be split into two phases: segment file selection, based on
time range; and reader construction, where we actually open the files.
It can happen that a file is removed (e.g. compacted) between the first
and second phase. In this case, we'll get partial results, but we
shouldn't fatal-out.

A better solution would be to open file handles when we select segments,
rather than just returning file names. I'll try to do that next.

* pkg/store: selecting matching segments opens them

This prevents a race between selecting segments by filename and opening
them as files, during which e.g. compaction can sneak in and delete the
files from the filesystem.

This reveals a lot of cleanup that should probably happen, at a minimum
around how and what we log.

* pkg/store: recover from bad filenames

Previously, any segment files that for whatever reason had bad filenames
(i.e. anything other than ULID-ULID.extension) could trigger a panic.
This change recovers from those situations more gracefully.

Also, store.FileLog read methods were using filepath.Walk instead of
fl.filesys.Walk -- oops! This presented no problem in the actual binary
but meant many tests weren't actually getting run :grimacing:

* pkg/store: use Event instead of raw log.Logger

A concrete type for events that can be emitted from package store.
By default, we just log these events via the logger. But, this commit
introduces some consistency for how they appear.

The hope is this will help to reveal the core problem that results in
invalid segment files.

* pkg/store: API: use EventReporter

* pkg/store: Compactor: use EventReporter

* pkg/store: Consumer: use EventReporter

* pkg/mmap: remove

The mmap filesys option was incorrectly implemented.

* pkg/ingest: improve test flake

Wait for the Write to be recorded before tearing down the conn.
Previously, the write may not have been fully flushed.

* pkg/store: fix bug re: recovering active files

.active files are the default extension for WriteSegments created by the
FileLog. They get UUID filenames because we don't yet know their low and
high ULIDs. Most of the time that's fine: code that expects to parse
filenames for ULID ranges (specifically: query code) knows to ignore
.active files.

When a FileLog is created, it first does a sweep of the underlying
filesystem, and any apparently temporary or intermediate files that are
discovered get recovered (essentially, renamed) to proper .flushed
files. This is a form of crash recovery. There was a bug in this code
that blindly changed the extension of .active files to .flushed, without
also giving them a proper low-high ULID filename. This causes a parse
failure, when query code tries to interpret the UUID as a pair of ULIDs.

This change fixes that, by extracting low and high ULIDs from .active
files, and renaming them properly during recovery.

* pkg/store: Parallelize more tests